\name{NoModelRun}
\alias{NoModelRun}
\title{
GFS model is not available
}
\description{
This function is called when \code{\link{GetModelRunHour}} tries to find a model that is not yet available on the NOMADS web site.
}
\usage{
NoModelRun(e)
}
\arguments{
  \item{e}{
  An error object.
}
}
\value{
INTERNAL
}
\author{
Daniel Bowman \email{daniel.bowman@unc.edu}
}
